VectorizedStatistics

Documentation for VectorizedStatistics.

VectorizedStatistics.vmaximum

VectorizedStatistics.vmean

VectorizedStatistics.vminimum

VectorizedStatistics.vsum

VectorizedStatistics.vmaximum — Method

vmaximum(A; dims)

As Base.maximum, but vectorized: find the greatest value contained in A, optionally over dimensions specified by dims.

Examples

julia> using VectorizedStatistics

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> vmaximum(A, dims=1)
1×2 Matrix{Int64}:
 3  4

julia>  vmaximum(A, dims=2)
 2×1 Matrix{Int64}:
 2
 4

VectorizedStatistics.vmean — Method

vmean(A; dims)

As Statistics.mean, but vectorized: compute the mean of all elements in A, optionally over dimensions specified by dims.

Examples

julia> using VectorizedStatistics

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> vmean(A, dims=1)
1×2 Matrix{Float64}:
 2.0  3.0

julia> vmean(A, dims=2)
2×1 Matrix{Float64}:
 1.5
 3.5

VectorizedStatistics.vminimum — Method

vminimum(A; dims)

As Base.minimum, but vectorized: find the least value contained in A, optionally over dimensions specified by dims.

Examples

julia> using VectorizedStatistics

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> vminimum(A, dims=1)
1×2 Matrix{Int64}:
 1  2

julia>  vminimum(A, dims=2)
 2×1 Matrix{Int64}:
 1
 3

VectorizedStatistics.vsum — Method

vsum(A; dims)

As Base.sum, but vectorized: summate the values contained in A, optionally over dimensions specified by dims.

Examples

julia> using VectorizedStatistics

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> vsum(A, dims=1)
1×2 Matrix{Int64}:
 4  6

julia> vsum(A, dims=2)
 2×1 Matrix{Int64}:
 3
 7